Externatilize templating
